partselect_chat/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.12.0
pgvector==0.2.4
django-cors-headers==4.3.1
python-decouple==3.8